Updated LupiTeam's domain
I wrote some days ago the server's IP address because someone keeps reporting our free domains. Now we bought one, so the problem should be solved. The server IP address is different and in future we will disable the older, so is worth change the domain. I hope this is the latest time.