LANraragi: Improve Random item visibility
Previously a server with a dirty DB would have conflicting numbers and need to run a clean to fix them.
That's still the case but that's not the extension's problem and those numbers are no longer considered.
Switched to a querySize check to remain only on the initial Browse page.
Remove previous Latest query smuggling due to above.