Fix Jingmantiantang's ownText is null and add deeplink support
[Fix] Jingmantiantang's ownText is null during chapter loading

The old implementation is to use chapterArea to store wheather or
not the Manga requested is only single page or multi-chapter.
This, however, creates a big issue as the chapterArea can be changed
to multi-chapter in preveious Manga, and that information is carried
over to other Manga which might be single chapter.
The new implementation remove global value chapterArea and it helper
function determineChapterInfo.
Instead, the single-multi check is performed in chapterListParse, which
only happen once per manga update request, vs three times call of
determineChapterInfo.

[Fix] Change Author's CSS selector and enable multi-author listing

[Fix] Default manga status to Completed

As Jingmantiantang often do not mark manga as complete, and there
is a high chance that manga user requested in single page. Thus
defaulting to completed, unless marked as ongoing.

[FEAT] Enable BaseUrl switch

As mentioned in (tachiyomiorg#5363) and (tachiyomiorg#4908). There is a frquent PR for
mirror site change. Unable user choice to stick with main site
or mirror site should please everyone

[FIX] Record mainSiteRateLimitPeriodPreference to correct key.

[FIX] Change one tag to the correct name

[FEAT] Enable deeplink support